--- app/src/main/java/com/bibbidi/habittracker/data/source/HabitsDataSource.kt
@@ -1,3 +1,0 @@
-package com.bibbidi.habittracker.data.source
-
-interface HabitsDataSource
