+++ app/src/main/java/com/bibbidi/habittracker/data/source/database/HabitsDao.kt
@@ -1,0 +1,6 @@
+package com.bibbidi.habittracker.data.source.database
+
+import androidx.room.Dao
+
+@Dao
+interface HabitsDao
