--- app/src/main/java/com/bibbidi/habittracker/data/source/local/HabitsDao.kt
@@ -1,6 +1,0 @@
-package com.bibbidi.habittracker.data.source.local
-
-import androidx.room.Dao
-
-@Dao
-interface HabitsDao
