--- app/src/main/java/com/bibbidi/habittracker/data/source/local/HabitsLocalDataSource.kt
@@ -1,8 +1,0 @@
-package com.bibbidi.habittracker.data.source.local
-
-import com.bibbidi.habittracker.data.source.HabitsDataSource
-import javax.inject.Inject
-
-class HabitsLocalDataSource @Inject constructor(
-    private val habitsDao: HabitsDao
-) : HabitsDataSource
